Prove that renders deleted comments
